+++ mlgidlab/__init__.py
@@ -15,7 +15,7 @@ os.environ.setdefault("HDF5_USE_FILE_LOCKING", "FALSE")
 
 from PySide6.QtWidgets import QApplication
 
-__version__ = "0.1.0a2"
+__version__ = "0.1.0a3"
 
 
 def main() -> int: